requirements.txt
bs4==0.0.2
dj-database-url==2.3.0
django-anymail==12.0
django-bootstrap5==24.3
django-cors-headers==4.6.0
django-pipeline==3.1.0
django-rq==3.0.0
django-taggit==6.1.0
django-watson==1.6.3
Django==5.1.2
easy-thumbnails==2.10
Gunicorn==23.0.0
markdown==3.7
notion-client==2.2.1
requests==2.32.3
